Advise sub-build systems to properly search the deps/ directory.
The --with-sysroot option to the `configure' script does not what was
intended. Autotools relies on gcc to find everything, which in turn
relies on the classical CFLAGS and LDFLAGS variables. Some projects
also use pkg-config, which can be configured with PKG_CONFIG_LIBDIR.

cmake-based projects are smart enough to pass CMAKE_PREFIX_PATH down
to pkg-config if it is used.