Implement proper path handling.
This introduces the OS namespace for functions specific to certain
operating systems. In it, defined currently are functions fo

  a) work with system-specific filesystem pathes,
  b) work with system-specific character encodings.

As it turns out, the constructor for std::filesystem::path does a good
job on dealing with the native filesystem encoding so the need for the
encoding transformation functions is rather low here. However, it is
assumed that at some point calls to e.g. the Win32API will be
required, in which case it will be handy to have these conversion
functions available.