Enable Ogre's DotSceneLoader plugin. Adds dependency on pugixml.
The DotSceneLoader plugin allows to load entire areas from XML files.
The blender2ogre plugin for Blender supports exporting to this XML
format from within Blender. Enabling this plugin is thus required for
the creation of entire areas in Blender. Without it, all objects in a
scene would have to be loaded (and positioned!) individually.