Add replaceBlenderEntities() for exchanging active objects.
In Blender, add a custom property `il_entity' to any object in order
to make it an active object. This function will find it in the scene
node tree and... print a message to the console that it found it. Next
step is to actually have this function replace the entity in question
with the real object requested by the designer with the `il_entity'
property.