Use Ogre's default axes layout.
There's no point in fighting the defaults all the time. Just use Y as
the upwards axes. When exporting from Blender, be sure to select the
option to swap the Y and Z axes, though.

As it turns out, Bullet uses the same axes orientation as Ogre, so
keeping the Z axis as upwards would have created even more hassle.