Use BtOgre for connecting Ogre with Bullet.
Rather than rolling our own implementation, use BtOgre and built it as
part of the normal build. This is because it turns out to be
surprisingly complicated to extract the original mesh vertices from a
loaded .mesh file from within Ogre; BtOgre includes code to do just
that. From a quick glance, it looks as if it is using several
undocumented interfaces to do that, but since BtOgre has achieved a
recognised status in Ogre (by being included into the OgreCave GitHub
account) this should be safe to use, but it’s not something to code
oneself. The alternative would have been to export Bullet geometry
(.bullet files) from Blender just like one exports .mesh files for
Ogre; this is a little convuluted and requires abusing Blender's own
game engine, but it is possible (see
<https://xissburg.com/post/export-bullet-from-blender/>, for an old
version of Blender [before 2.8]). The approach taken here is the
approach of least resistance.

It looks as if BtOgre needs some adaptions to fit into the Ilmendur
architecture as I imagine it (there’s no removeRigidBody() function
for instance), so include it rather than referencing it as a CMake
ExternalProject. Also, BtOgre does not have formal releases anyway, so
it’s more sensible to include it directly into the source tree.

This commit, when compiled and run, will show the effects of a physics
collision of an object falling down onto the test area ground.