Fix incorrectly calculated bounding box.
CT_BOX collision boxes require that the model's origin is actually in
the mesh’s geometrical centre, otherwise the collision box is set off.
It seems that BtOgre calculates a collision box by taking the distance
from the object origin to the furthest vertex, multiplying it by two
and completing the remaining edges.