Fix joystick detection.
It turns out some joystick axes are at -1 or 1 if not touched. So
instead of trying to find the axis with the maximum absolute value,
the code now queries the neutral position and then afterwards
instructs the user to touch the joystick. From the difference the
respective axis is determined.