Clean up physics enablement API.
The previous approach required subclass writers to set protected
member variables from the Actor class in the subclass in the correct
order before manually instancing the RigidBody class, for which then
existed seemingly no corresponding `delete' statement (because it was
in Actor’s destructor). The new approach is cleaner. It uses a
protected method Actor::addRigidBody() which does all things in the
correct order and hides the memory management of the RigidBody
instance from subclass writers. This resulted in the corresponding
member variables being made `private'; subclass writers now access
them through the ordinary getter methods.