Introduce activate/deactivtate on Scene objects.
Ogre has some global resources -- namely the render window -- which
must not be acquired by two scenes at the same time, otherwise Ogre
will crash. This commit moves the acquisiation of such global
resources from the respective scenes’ constructors into a new callback
function activate(), which is called when the respective scene makes
it to the top of the scene stack. An inverse deactivate() is provided
as well.

The changes in ui.cpp cater for the fact that a similar problem exists
with the Ogre UI overlay. It must only be instanciated once, otherwise
it will try to load the font resources again, which will cause Ogre to
crash with a duplicate font resource error.