Complement the camera view with a cliprect.
This is because for whatever reason, SDL does not cut off drawing
material beyond the viewport's width and height if the viewport is
located at (0|0), at least on my Linux OpenGL system. It does cut off
things if the viewport is located elsewhere. Clipping however always
properly cuts off, it just needs to be taken into account that
clipping rectangles are relative to viewports. And that for whatever
reason, it is required to clear the clipping rectangle before setting
a new one, otherwise nothing will be shown.