More dependency fiddling.
So, SDL does not normally use the Linux runtime linker, but wants to
load all of its dynamic dependencies with dlopen(3). This has the
unfortunate consequence that it is not obvious for which of SDL's
dependencies Ilmendur's CMakeLists.txt has to test, because the build
will simply not fail at compile time, but only cause runtime errors.
This commit reflects what I found by experimenting with SDL_LOADSO=OFF
passed to SDL's build configuration, which forces SDL to not use
dlopen(3). Any missing library will cause a proper compilation error
in that case. But unfortunately, using that compilation option seems
to hit a bug in SDL's Wayland support as SDL will crash on Wayland at
runtime if that option was passed; it will however run on X11,
including on XWayland.