Add Flatpak build manifest.
It is a little hacky as flatpak-builder refuses network access during
the build process, which Ilmendur's build process does rely on,
because it uses CMake's ExternalProject command to download dependency
source files. The trick employed here is that it is simply necessary
to once build Ilmendur on the host system as usual (albeit in a
directory called "build") before running flatpak-builder to create a
flatpak. This way the sources have already been downloaded. The
Flatpak manifest includes a source specification of type "shell" which
does nothing else than to copy the already downloaded files where the
in-flatpak-builder cmake expects them. The in-flatpak-builder-cmake
will recognise the files are already there and skip the download step.

The downside of this is that creating a flatpak will always compile
the entire dependency tree, but at least in this early stage of
development, this appears to be fine. On the positive side this allows
to keep the dependency listing centralised at a single
place (cmake/dependencies.cmake).