Make the Flatpak build a little cleaner.
Previously, it was necessary to successfully run cmake on the host
system in a regular build process so far that it has downloaded all
source tarballs. Since the Flatpak build only requires the source
tarballs, this was a little over the top. Instead, this commit now
introduces a maintenance Rakefile which simply extracts what to
download from cmake/dependencies.cmake, and downloads that when the
creation of the Flatpak bundle is demanded with the new
ilmendur.flatpak file task.

That is, in order to build the flatpak bundle, do not execute
`flatpak-builder' or `flatpak bundle directly'. Instead, execute this:

    $ rake ilmendur.flatpak