Promote map layers to their own proper class.
This commit refactors the relationship between the Actor, Map, and
layer-related classes. Previously, Map owned all layers and managed
the actors by themselves. This became unwieldy, as it turned out to be
quite frequently necessary to go from the Actor instance backward to
the Map, which before this commit required an expensive search
operation implemented in Map::findActor(). The need for such searching
is now removed, and the resulting API provides for a much cleaner
model of ownership. The new model is simply:

  Map ----> MapLayer subclasses ----> Actor
      owns                       own

With backward-links being provided in MapLayer and Actor.