Allow translation of text strings in TMX files.
The idea is that the i18n:potfile rake task (via a dependency task)
extracts those strings from the TMX files and places them in a no-op
file external_strings.cpp. This file in turn is processed by
xgettext(1) for the translator's work sheet (PO file). At runtime, the
game loads the untranslated string from the TMX file and then passes
that one to gettext() unmodified; the string is found in the catalogue
still (if it has been translated), because it was included in the PO
file due to the trick with external_strings.cpp.

external_strings.cpp is generated, but only occasionally when the POT
file is updated. Like the POT file, it should thus be in the VCS.